+++ ACNHMS_Source/Assets/Scripts/Injection/SysBot/USBBot.cs
@@ -159,7 +159,7 @@ namespace NHSE.Injection
             }
         }
 
-        public byte[] ReadBytes(uint offset, int length, RWMethod method = RWMethod.Heap)
+        public byte[] ReadBytes(ulong offset, int length, RWMethod method = RWMethod.Heap)
         {
             if (length > MaximumTransferSize)
                 return ReadBytesLarge(offset, length, method);
@@ -178,7 +178,7 @@ namespace NHSE.Injection
             }
         }
 
-        public void WriteBytes(byte[] data, uint offset, RWMethod method = RWMethod.Heap)
+        public void WriteBytes(byte[] data, ulong offset, RWMethod method = RWMethod.Heap)
         {
             if (data.Length > MaximumTransferSize)
                 WriteBytesLarge(data, offset, method);
@@ -206,14 +206,14 @@ namespace NHSE.Injection
             }
         }
 
-        private void WriteBytesLarge(byte[] data, uint offset, RWMethod method)
+        private void WriteBytesLarge(byte[] data, ulong offset, RWMethod method)
         {
             int byteCount = data.Length;
             for (int i = 0; i < byteCount; i += MaximumTransferSize)
                 WriteBytes(SubArray(data, i, MaximumTransferSize), offset + (uint)i, method);
         }
 
-        private byte[] ReadBytesLarge(uint offset, int length, RWMethod method)
+        private byte[] ReadBytesLarge(ulong offset, int length, RWMethod method)
         {
             List<byte> read = new List<byte>();
             for (int i = 0; i < length; i += MaximumTransferSize)
